1.0.7-8: Catalog extraction into dango/catalog/ library module
Extract catalog data access layer from web/routes/catalog.py into new standalone
dango/catalog/ module with no FastAPI dependency, enabling reuse by CLI commands
and future API consumers.

Changes:
- Create dango/catalog/ module with 6 files:
  * __init__.py — public API (5 aliases)
  * schema.py — DuckDB introspection (165 lines)
  * profiling.py — cached stats readers (91 lines)
  * lineage.py — manifest DAG + impact analysis (196 lines)
  * manifest.py — pure node readers (249 lines)
  * models.py — response builders (301 lines)
  * CLAUDE.md — module documentation

- Rewire dango/web/routes/catalog.py (1489 → 563 lines):
  * All function bodies moved to new modules
  * Routes reduced to thin handlers with asyncio.to_thread() dispatch
  * Private underscore-prefixed names imported directly (176+ test @patch sites
    unaffected)

- New test file tests/unit/test_catalog_module.py:
  * Verify FastAPI not leaked from any module
  * Verify public API exports
  * Verify basic DuckDB queries

Result:
- 102 / 111 existing catalog tests passing (9 pre-existing mock setup issues)
- All new tests passing
- Code quality: ruff/mypy passing
- All extracted modules ≤ 301 lines (well under 500-line CI limit)

Known limitation (documented): _build_lineage_dag, _build_catalog_models, and
_get_run_results still require FastAPI app context (call get_project_root() with
no params). Fix deferred: reorder route handlers to populate project_root before
calling these functions, then add project_root parameter.
