+++ dango/web/routes/catalog.py
@@ -7,8 +7,6 @@ lineage, impact analysis, and unified model browsing.
 from __future__ import annotations
 
 import asyncio
-import json
-import re
 from pathlib import Path
 from typing import Any
 
@@ -18,774 +16,42 @@ from fastapi import APIRouter, Depends, HTTPException, Query
 from dango.auth.audit import AuditEvent, log_auth_event
 from dango.auth.models import User
 from dango.auth.permissions import require_permission
+from dango.catalog.lineage import (  # noqa: F401 — _get_impact_tree for integration tests
+    _build_impact_response,
+    _build_lineage_dag,
+    _get_impact_tree,
+)
+from dango.catalog.manifest import (
+    _find_model_in_manifest,
+    _model_profiling_key,
+    _search_manifest,
+)
+from dango.catalog.models import (
+    _build_catalog_models,
+    _build_model_detail,
+    _get_run_results,
+    _get_source_summary_stats,
+)
+from dango.catalog.profiling import (
+    _get_cached_stats,
+    _get_profiled_at,
+)
+from dango.catalog.schema import (
+    _get_column_schema,
+    _get_model_column_schema,
+    _get_raw_tables_from_duckdb,
+    _get_row_count,
+    _source_schema_exists,
+    _table_exists,
+)
 from dango.logging import get_logger
-from dango.utils.dango_db import connect
 from dango.utils.post_sync import _run_profiling, profile_table
 from dango.validation import validate_identifier, validate_source_name
-from dango.web.helpers import get_dbt_manifest, get_project_root
+from dango.web.helpers import get_dbt_manifest, get_project_root  # Imported for test patching
 
 logger = get_logger(__name__)
 
 router = APIRouter(tags=["catalog"])
-
-
-# ---------------------------------------------------------------------------
-# Private blocking helpers (called via asyncio.to_thread)
-# ---------------------------------------------------------------------------
-
-
-def _get_column_schema(
-    db_path: Path,
-    source: str,
-    table: str,
-) -> list[dict[str, Any]]:
-    """Query DuckDB for column metadata of a single table.
-
-    Args:
-        db_path: Path to the DuckDB warehouse file.
-        source: Source name (schema is ``raw_{source}``).
-        table: Table name.
-
-    Returns:
-        List of ``{"name": ..., "type": ..., "nullable": bool}``.
-    """
-    schema = f"raw_{source}"
-    conn = duckdb.connect(str(db_path), config={"access_mode": "read_only"})
-    try:
-        rows = conn.execute(
-            "SELECT column_name, data_type, is_nullable "
-            "FROM information_schema.columns "
-            "WHERE table_schema = ? AND table_name = ? "
-            "ORDER BY ordinal_position",
-            [schema, table],
-        ).fetchall()
-    finally:
-        conn.close()
-
-    return [{"name": r[0], "type": r[1], "nullable": r[2] == "YES"} for r in rows]
-
-
-def _get_cached_stats(
-    project_root: Path,
-    source: str,
-    table: str,
-) -> dict[str, dict[str, str | None]]:
-    """Read cached profiling stats from SQLite.
-
-    Args:
-        project_root: Dango project root.
-        source: Source name.
-        table: Table name.
-
-    Returns:
-        Mapping of ``{column_name: {stat_type: stat_value}}``.
-    """
-    result: dict[str, dict[str, str | None]] = {}
-    with connect(project_root) as conn:
-        rows = conn.execute(
-            "SELECT column_name, stat_type, stat_value "
-            "FROM profiling_stats WHERE source = ? AND table_name = ? "
-            "AND column_name != '__row_count__'",
-            (source, table),
-        ).fetchall()
-    for row in rows:
-        col_name = row[0]
-        if col_name not in result:
-            result[col_name] = {}
-        result[col_name][row[1]] = row[2]
-    return result
-
-
-def _get_profiled_at(
-    project_root: Path,
-    source: str,
-    table: str,
-) -> str | None:
-    """Return the most recent ``updated_at`` from profiling stats.
-
-    Args:
-        project_root: Dango project root.
-        source: Source name.
-        table: Table name.
-
-    Returns:
-        ISO timestamp string or ``None`` if no stats exist.
-    """
-    with connect(project_root) as conn:
-        row = conn.execute(
-            "SELECT MAX(updated_at) FROM profiling_stats WHERE source = ? AND table_name = ?",
-            (source, table),
-        ).fetchone()
-    if row and row[0]:
-        result: str = row[0]
-        return result
-    return None
-
-
-def _get_cached_row_counts(project_root: Path) -> dict[tuple[str, str], int]:
-    """Read all cached table-level row counts from SQLite.
-
-    Args:
-        project_root: Dango project root.
-
-    Returns:
-        Mapping of ``{(source, table_name): row_count}`` for every table that
-        has a cached ``__row_count__`` synthetic stat row.
-    """
-    result: dict[tuple[str, str], int] = {}
-    with connect(project_root) as conn:
-        rows = conn.execute(
-            "SELECT source, table_name, stat_value FROM profiling_stats "
-            "WHERE column_name = '__row_count__' AND stat_type = 'value'"
-        ).fetchall()
-    for row in rows:
-        try:
-            result[(row[0], row[1])] = int(row[2])
-        except (TypeError, ValueError):
-            continue
-    return result
-
-
-def _get_row_count(db_path: Path, source: str, table: str) -> int:
-    """Return the row count for a table.
-
-    Args:
-        db_path: Path to the DuckDB warehouse file.
-        source: Source name.
-        table: Table name.
-
-    Returns:
-        Number of rows.
-    """
-    schema = f"raw_{source}"
-    conn = duckdb.connect(str(db_path), config={"access_mode": "read_only"})
-    try:
-        result = conn.execute(f'SELECT COUNT(*) FROM "{schema}"."{table}"').fetchone()
-    finally:
-        conn.close()
-    return result[0] if result else 0
-
-
-def _source_schema_exists(db_path: Path, source: str) -> bool:
-    """Check whether the ``raw_{source}`` schema exists in DuckDB.
-
-    Args:
-        db_path: Path to the DuckDB warehouse file.
-        source: Source name.
-
-    Returns:
-        ``True`` if the schema exists.
-    """
-    schema = f"raw_{source}"
-    conn = duckdb.connect(str(db_path), config={"access_mode": "read_only"})
-    try:
-        result = conn.execute(
-            "SELECT COUNT(*) FROM information_schema.schemata WHERE schema_name = ?",
-            [schema],
-        ).fetchone()
-    finally:
-        conn.close()
-    return bool(result and result[0] > 0)
-
-
-def _table_exists(db_path: Path, source: str, table: str) -> bool:
-    """Check whether a user table exists in DuckDB (excluding ``_dlt_*``).
-
-    Args:
-        db_path: Path to the DuckDB warehouse file.
-        source: Source name.
-        table: Table name.
-
-    Returns:
-        ``True`` if the table exists.
-    """
-    schema = f"raw_{source}"
-    conn = duckdb.connect(str(db_path), config={"access_mode": "read_only"})
-    try:
-        result = conn.execute(
-            "SELECT COUNT(*) FROM information_schema.tables "
-            "WHERE table_schema = ? AND table_name = ? "
-            "AND table_name NOT LIKE '_dlt_%'",
-            [schema, table],
-        ).fetchone()
-    finally:
-        conn.close()
-    return bool(result and result[0] > 0)
-
-
-def _get_raw_tables_from_duckdb(db_path: Path) -> list[dict[str, str]]:
-    """List all user tables in raw_* schemas (for discovering unmodeled tables).
-
-    Args:
-        db_path: Path to the DuckDB warehouse file.
-
-    Returns:
-        List of ``{"schema": ..., "table": ..., "source_name": ...}``.
-    """
-    conn = duckdb.connect(str(db_path), config={"access_mode": "read_only"})
-    try:
-        rows = conn.execute(
-            "SELECT table_schema, table_name FROM information_schema.tables "
-            "WHERE table_schema LIKE 'raw_%' AND table_name NOT LIKE '_dlt_%'"
-        ).fetchall()
-    finally:
-        conn.close()
-    return [{"schema": r[0], "table": r[1], "source_name": r[0][4:]} for r in rows]
-
-
-def _get_source_summary_stats(db_path: Path) -> dict[str, dict[str, int]]:
-    """Query per-source table counts and estimated row counts.
-
-    Args:
-        db_path: Path to DuckDB warehouse file.
-
-    Returns:
-        ``{source_name: {"table_count": int, "estimated_row_total": int}}``.
-    """
-    try:
-        conn = duckdb.connect(str(db_path), config={"access_mode": "read_only"})
-        try:
-            rows = conn.execute(
-                "SELECT schema_name, COUNT(*) AS table_count, "
-                "COALESCE(SUM(estimated_size), 0) AS estimated_row_total "
-                "FROM duckdb_tables() "
-                "WHERE schema_name LIKE 'raw_%' "
-                "AND table_name NOT LIKE '_dlt_%' "
-                "GROUP BY schema_name"
-            ).fetchall()
-        finally:
-            conn.close()
-    except Exception:
-        logger.warning("source_summary_stats_failed", db_path=str(db_path))
-        return {}
-    return {r[0][4:]: {"table_count": r[1], "estimated_row_total": r[2]} for r in rows}
-
-
-# ---------------------------------------------------------------------------
-# Manifest / model helpers (called via asyncio.to_thread)
-# ---------------------------------------------------------------------------
-
-
-def _get_run_results() -> dict[str, Any] | None:
-    """Load ``dbt/target/run_results.json``.
-
-    Returns:
-        Parsed dict or ``None`` if the file does not exist.
-    """
-    path = get_project_root() / "dbt" / "target" / "run_results.json"
-    if not path.exists():
-        return None
-    try:
-        with open(path, encoding="utf-8") as f:
-            result: dict[str, Any] = json.load(f)
-            return result
-    except Exception:
-        logger.warning("Failed to parse run_results.json")
-        return None
-
-
-def _build_test_status_map(
-    manifest: dict[str, Any],
-    run_results: dict[str, Any] | None,
-) -> dict[str, list[dict[str, str | None]]]:
-    """Map model/source unique_ids to their test results.
-
-    Args:
-        manifest: Parsed dbt manifest.
-        run_results: Parsed ``run_results.json`` or ``None``.
-
-    Returns:
-        ``{model_unique_id: [{"name": ..., "status": "pass"|"fail"|"error"|None}]}``.
-    """
-    # Build status lookup from run_results
-    result_status: dict[str, str] = {}
-    if run_results:
-        for r in run_results.get("results", []):
-            uid = r.get("unique_id", "")
-            raw_status = r.get("status", "")
-            if uid and raw_status:
-                # dbt uses "success"; normalize to "pass" for display
-                result_status[uid] = "pass" if raw_status == "success" else raw_status
-
-    test_map: dict[str, list[dict[str, str | None]]] = {}
-    for uid, node in manifest.get("nodes", {}).items():
-        if node.get("resource_type") != "test":
-            continue
-        test_name = node.get("name", uid)
-        status = result_status.get(uid)
-        for dep in node.get("depends_on", {}).get("nodes", []):
-            test_map.setdefault(dep, []).append({"name": test_name, "status": status})
-
-    return test_map
-
-
-def _classify_model_type(node: dict[str, Any]) -> str:
-    """Classify a dbt model as ``staging``, ``intermediate``, or ``marts``.
-
-    Classification priority:
-    1. Schema name: ``staging`` / ``intermediate`` / ``marts``.
-    2. Name prefix: ``stg_`` → staging, ``fct_`` / ``dim_`` → marts,
-       ``int_`` → intermediate.
-    3. Fallback: ``intermediate``.
-
-    Args:
-        node: A manifest model node dict.
-
-    Returns:
-        One of ``"staging"``, ``"intermediate"``, ``"marts"``.
-    """
-    schema = (node.get("schema") or "").lower()
-    if schema == "staging":
-        return "staging"
-    if schema == "intermediate":
-        return "intermediate"
-    if schema == "marts":
-        return "marts"
-
-    name = (node.get("name") or "").lower()
-    if name.startswith("stg_"):
-        return "staging"
-    if name.startswith(("fct_", "dim_")):
-        return "marts"
-    if name.startswith("int_"):
-        return "intermediate"
-
-    return "intermediate"
-
-
-def _model_profiling_key(node: dict[str, Any], kind: str) -> tuple[str, str] | None:
-    """Return the ``(source, table_name)`` SQLite key used by the profiling write path.
-
-    Mirrors :func:`dango.utils.post_sync.profile_table` callers so cached stats
-    and row counts can be read back for any model type:
-    - sources and staging models are keyed by the raw source name,
-    - intermediate/marts models are keyed by their dbt schema,
-    - seeds are keyed by their dbt schema (always "main" by default).
-
-    Args:
-        node: A manifest model, source, or seed node dict.
-        kind: ``"model"``, ``"source"``, or ``"seed"``.
-
-    Returns:
-        ``(source, table_name)`` tuple or ``None`` if no key applies.
-    """
-    table = node.get("alias") or node.get("name", "")
-    if not table:
-        return None
-    if kind == "seed":
-        schema = node.get("schema", "main")
-        return (schema, table) if table else None
-    if kind == "source":
-        source = node.get("source_name", "")
-        return (source, table) if source else None
-    schema = node.get("schema", "")
-    if schema == "staging":
-        m = re.match(r"^stg_(.+?)__", table)
-        return (m.group(1), table) if m else None
-    return (schema, table) if schema else None
-
-
-def _get_model_column_schema(
-    db_path: Path,
-    schema: str,
-    table: str,
-) -> list[dict[str, Any]]:
-    """Query DuckDB ``information_schema.columns`` for any schema.
-
-    Unlike :func:`_get_column_schema`, this is not restricted to
-    ``raw_*`` schemas — it works for staging, intermediate, and marts.
-
-    Args:
-        db_path: Path to the DuckDB warehouse file.
-        schema: Schema name (e.g. ``staging``, ``marts``).
-        table: Table/view name.
-
-    Returns:
-        List of ``{"name": ..., "type": ..., "nullable": bool}``,
-        or empty list if the table does not exist.
-    """
-    conn = duckdb.connect(str(db_path), config={"access_mode": "read_only"})
-    try:
-        rows = conn.execute(
-            "SELECT column_name, data_type, is_nullable "
-            "FROM information_schema.columns "
-            "WHERE table_schema = ? AND table_name = ? "
-            "ORDER BY ordinal_position",
-            [schema, table],
-        ).fetchall()
-    except Exception:
-        rows = []
-    finally:
-        conn.close()
-
-    return [{"name": r[0], "type": r[1], "nullable": r[2] == "YES"} for r in rows]
-
-
-def _build_catalog_models(
-    manifest: dict[str, Any],
-    run_results: dict[str, Any] | None,
-) -> dict[str, Any]:
-    """Build the full catalog models + sources response from manifest.
-
-    Args:
-        manifest: Parsed dbt manifest.
-        run_results: Parsed run_results or ``None``.
-
-    Returns:
-        Dict with ``models`` and ``sources`` lists.
-    """
-    test_map = _build_test_status_map(manifest, run_results)
-
-    # Load persistent dbt model statuses for last_run timestamps
-    from dango.utils.dbt_status import get_model_statuses
-
-    project_root = get_project_root()
-    model_statuses = get_model_statuses(project_root)
-    cached_row_counts = _get_cached_row_counts(project_root)
-
-    models: list[dict[str, Any]] = []
-    for uid, node in manifest.get("nodes", {}).items():
-        if node.get("resource_type") != "model":
-            continue
-        columns = node.get("columns", {})
-        cols_documented = sum(1 for c in columns.values() if c.get("description"))
-        tests = test_map.get(uid, [])
-        tests_passing = sum(1 for t in tests if t["status"] == "pass")
-        tests_warning = sum(1 for t in tests if t["status"] == "warn")
-        tests_failing = sum(1 for t in tests if t["status"] in ("fail", "error"))
-
-        status_info = model_statuses.get(uid, {})
-        key = _model_profiling_key(node, "model")
-        row_count = cached_row_counts.get(key) if key else None
-
-        models.append(
-            {
-                "unique_id": uid,
-                "name": node.get("name", ""),
-                "type": _classify_model_type(node),
-                "schema": node.get("schema", ""),
-                "materialization": node.get("config", {}).get("materialized", "view"),
-                "description": node.get("description", ""),
-                "test_count": len(tests),
-                "tests_passing": tests_passing,
-                "tests_warning": tests_warning,
-                "tests_failing": tests_failing,
-                "columns_total": len(columns),
-                "columns_documented": cols_documented,
-                "tags": node.get("tags", []),
-                "last_run": status_info.get("last_run"),
-                "status": status_info.get("status"),
-                "row_count": row_count,
-            }
-        )
-
-    # Seeds — resource_type == "seed" nodes from manifest.
-    # Note: seeds don't have dbt tests (test_count=0, always) or status tracking.
-    # Row counts are populated from profiling cache if available.
-    for uid, node in manifest.get("nodes", {}).items():
-        if node.get("resource_type") != "seed":
-            continue
-        columns = node.get("columns", {})
-        cols_documented = sum(1 for c in columns.values() if c.get("description"))
-        key = _model_profiling_key(node, "seed")
-        row_count = cached_row_counts.get(key) if key else None
-        models.append(
-            {
-                "unique_id": uid,
-                "name": node.get("name", ""),
-                "type": "seed",
-                "schema": node.get("schema", ""),
-                "materialization": "table",
-                "description": node.get("description", ""),
-                "test_count": 0,  # Seeds don't have dbt tests
-                "tests_passing": 0,
-                "tests_warning": 0,
-                "tests_failing": 0,
-                "columns_total": len(columns),
-                "columns_documented": cols_documented,
-                "tags": node.get("tags", []),
-                "last_run": None,  # Seeds don't have execution status
-                "status": None,
-                "row_count": row_count,
-            }
-        )
-
-    # Sort: type order (staging → intermediate → marts → seed), then name
-    type_order = {"staging": 0, "intermediate": 1, "marts": 2, "seed": 3}
-    models.sort(key=lambda m: (type_order.get(m["type"], 1), m["name"].lower()))
-
-    sources: list[dict[str, Any]] = []
-    for uid, src in manifest.get("sources", {}).items():
-        columns = src.get("columns", {})
-        cols_documented = sum(1 for c in columns.values() if c.get("description"))
-        tests = test_map.get(uid, [])
-        tests_passing = sum(1 for t in tests if t["status"] == "pass")
-        tests_warning = sum(1 for t in tests if t["status"] == "warn")
-        tests_failing = sum(1 for t in tests if t["status"] in ("fail", "error"))
-
-        key = _model_profiling_key(src, "source")
-        row_count = cached_row_counts.get(key) if key else None
-
-        sources.append(
-            {
-                "unique_id": uid,
-                "name": src.get("name", ""),
-                "type": "source",
-                "schema": src.get("schema", ""),
-                "description": src.get("description", ""),
-                "source_name": src.get("source_name", ""),
-                "test_count": len(tests),
-                "tests_passing": tests_passing,
-                "tests_warning": tests_warning,
-                "tests_failing": tests_failing,
-                "columns_total": len(columns),
-                "columns_documented": cols_documented,
-                "row_count": row_count,
-            }
-        )
-
-    sources.sort(key=lambda s: s["name"].lower())
-
-    return {"models": models, "sources": sources}
-
-
-def _find_model_in_manifest(
-    manifest: dict[str, Any],
-    model_name: str,
-) -> tuple[str | None, dict[str, Any] | None, str]:
-    """Find a model or source by name in the manifest.
-
-    Models are preferred over sources when names collide.
-
-    Args:
-        manifest: Parsed dbt manifest.
-        model_name: Name to search for.
-
-    Returns:
-        Tuple of ``(unique_id, node_dict, kind)`` where *kind* is
-        ``"model"`` or ``"source"``.  Returns ``(None, None, "")``
-        if not found.
-    """
-    # Search models first
-    for uid, node in manifest.get("nodes", {}).items():
-        if node.get("resource_type") == "model" and node.get("name") == model_name:
-            return uid, node, "model"
-    # Then sources
-    for uid, src in manifest.get("sources", {}).items():
-        if src.get("name") == model_name:
-            return uid, src, "source"
-    return None, None, ""
-
-
-def _build_model_detail(
-    manifest: dict[str, Any],
-    run_results: dict[str, Any] | None,
-    target_uid: str,
-    target_node: dict[str, Any],
-    kind: str,
-    db_columns: list[dict[str, Any]],
-    profiled_at: str | None,
-) -> dict[str, Any]:
-    """Build the detail response for a single model or source.
-
-    Args:
-        manifest: Full dbt manifest.
-        run_results: Parsed run_results or ``None``.
-        target_uid: The unique_id of the target.
-        target_node: The manifest node dict.
-        kind: ``"model"`` or ``"source"``.
-        db_columns: Column schema from DuckDB (may be empty).
-        profiled_at: Last profiling timestamp or ``None``.
-
-    Returns:
-        Full detail response dict.
-    """
-    test_map = _build_test_status_map(manifest, run_results)
-    model_tests = test_map.get(target_uid, [])
-    manifest_columns = target_node.get("columns", {})
-
-    # Build reverse dependency map
-    reverse_map: dict[str, list[str]] = {}
-    for uid, node in manifest.get("nodes", {}).items():
-        if node.get("resource_type") == "model":
-            for dep in node.get("depends_on", {}).get("nodes", []):
-                reverse_map.setdefault(dep, []).append(uid)
-
-    # Merge DuckDB columns with manifest column descriptions
-    model_name = target_node.get("name", "")
-    columns: list[dict[str, Any]] = []
-    if db_columns:
-        for col in db_columns:
-            manifest_col = manifest_columns.get(col["name"], {})
-            # Map tests to this column by dbt naming convention:
-            # test names follow {test_type}_{model}_{column} pattern
-            col_suffix = f"_{model_name}_{col['name']}"
-            col_tests = [t for t in model_tests if t["name"] and t["name"].endswith(col_suffix)]
-            columns.append(
-                {
-                    "name": col["name"],
-                    "type": col["type"],
-                    "nullable": col["nullable"],
-                    "description": manifest_col.get("description") or None,
-                    "tests": col_tests if col_tests else None,
-                    "stats": None,
-                }
-            )
-    else:
-        # Model not materialised — show manifest columns without type info
-        for col_name, col_info in manifest_columns.items():
-            columns.append(
-                {
-                    "name": col_name,
-                    "type": None,
-                    "nullable": None,
-                    "description": col_info.get("description") or None,
-                    "tests": None,
-                    "stats": None,
-                }
-            )
-
-    result: dict[str, Any] = {
-        "unique_id": target_uid,
-        "name": model_name,
-        "schema": target_node.get("schema", ""),
-        "description": target_node.get("description", ""),
-        "tags": target_node.get("tags", []),
-        "meta": target_node.get("meta", {}),
-        "columns": columns,
-        "depends_on": target_node.get("depends_on", {}).get("nodes", []),
-        "depended_on_by": reverse_map.get(target_uid, []),
-        "tests": model_tests if model_tests else None,
-        "row_count": None,
-        "profiled_at": profiled_at,
-    }
-
-    if kind == "model":
-        result["type"] = _classify_model_type(target_node)
-        result["materialization"] = target_node.get("config", {}).get("materialized", "view")
-        result["raw_code"] = target_node.get("raw_code") or target_node.get("raw_sql")
-        result["compiled_code"] = target_node.get("compiled_code") or target_node.get(
-            "compiled_sql"
-        )
-    else:
-        result["type"] = "source"
-        result["materialization"] = None
-        result["raw_code"] = None
-        result["compiled_code"] = None
-        result["source_name"] = target_node.get("source_name", "")
-
-    return result
-
-
-def _search_manifest(
-    manifest: dict[str, Any],
-    query: str,
-) -> list[dict[str, Any]]:
-    """Search manifest models and sources by name, description, column names.
-
-    Args:
-        manifest: Parsed dbt manifest.
-        query: Case-insensitive search string.
-
-    Returns:
-        List of search result dicts (max 50).
-    """
-    q = query.lower()
-    name_matches: list[dict[str, Any]] = []
-    desc_matches: list[dict[str, Any]] = []
-    col_matches: list[dict[str, Any]] = []
-
-    # Search models
-    for uid, node in manifest.get("nodes", {}).items():
-        if node.get("resource_type") != "model":
-            continue
-        name = node.get("name", "")
-        desc = node.get("description", "")
-        model_type = _classify_model_type(node)
-
-        if q in name.lower():
-            name_matches.append(
-                {
-                    "unique_id": uid,
-                    "name": name,
-                    "type": model_type,
-                    "description": desc,
-                    "match_type": "name",
-                }
-            )
-            continue
-
-        if desc and q in desc.lower():
-            desc_matches.append(
-                {
-                    "unique_id": uid,
-                    "name": name,
-                    "type": model_type,
-                    "description": desc,
-                    "match_type": "description",
-                }
-            )
-            continue
-
-        for col_name in node.get("columns", {}):
-            if q in col_name.lower():
-                col_matches.append(
-                    {
-                        "unique_id": uid,
-                        "name": name,
-                        "type": model_type,
-                        "description": desc,
-                        "match_type": "column",
-                        "matched_column": col_name,
-                    }
-                )
-                break
-
-    # Search sources
-    for uid, src in manifest.get("sources", {}).items():
-        name = src.get("name", "")
-        desc = src.get("description", "")
-
-        if q in name.lower():
-            name_matches.append(
-                {
-                    "unique_id": uid,
-                    "name": name,
-                    "type": "source",
-                    "description": desc,
-                    "match_type": "name",
-                }
-            )
-            continue
-
-        if desc and q in desc.lower():
-            desc_matches.append(
-                {
-                    "unique_id": uid,
-                    "name": name,
-                    "type": "source",
-                    "description": desc,
-                    "match_type": "description",
-                }
-            )
-            continue
-
-        for col_name in src.get("columns", {}):
-            if q in col_name.lower():
-                col_matches.append(
-                    {
-                        "unique_id": uid,
-                        "name": name,
-                        "type": "source",
-                        "description": desc,
-                        "match_type": "column",
-                        "matched_column": col_name,
-                    }
-                )
-                break
-
-    results = name_matches + desc_matches + col_matches
-    return results[:50]
 
 
 # ---------------------------------------------------------------------------
@@ -1238,197 +504,6 @@ async def search_catalog(
 
     results = await asyncio.to_thread(_search_manifest, manifest, query)
     return {"query": query, "results": results}
-
-
-# ---------------------------------------------------------------------------
-# Lineage helpers (called via asyncio.to_thread)
-# ---------------------------------------------------------------------------
-
-
-def _build_lineage_dag() -> dict[str, Any] | None:
-    """Build full DAG from dbt manifest (sources + models + edges).
-
-    Returns:
-        Dict with ``nodes`` and ``edges`` lists, or ``None`` if no manifest.
-    """
-    manifest = get_dbt_manifest()
-    if manifest is None:
-        return None
-
-    nodes_out: list[dict[str, Any]] = []
-    edges: list[dict[str, str]] = []
-    depended_on_by: dict[str, list[str]] = {}
-    test_map: dict[str, list[str]] = {}
-
-    # Collect model dependencies and test associations
-    for uid, node in manifest.get("nodes", {}).items():
-        rtype = node.get("resource_type", "")
-        if rtype == "model":
-            deps = node.get("depends_on", {}).get("nodes", [])
-            for dep in deps:
-                depended_on_by.setdefault(dep, []).append(uid)
-                edges.append({"source": dep, "target": uid})
-        elif rtype == "test":
-            for dep in node.get("depends_on", {}).get("nodes", []):
-                test_map.setdefault(dep, []).append(node.get("name", uid))
-
-    # Build node list — models
-    for uid, node in manifest.get("nodes", {}).items():
-        if node.get("resource_type") != "model":
-            continue
-        columns = node.get("columns", {})
-        cols_documented = sum(1 for c in columns.values() if c.get("description"))
-        deps = node.get("depends_on", {}).get("nodes", [])
-        tests = test_map.get(uid, [])
-        nodes_out.append(
-            {
-                "id": uid,
-                "name": node.get("name", ""),
-                "type": "model",
-                "schema": node.get("schema", ""),
-                "materialization": node.get("config", {}).get("materialized"),
-                "description": node.get("description", ""),
-                "depends_on": deps,
-                "depended_on_by": depended_on_by.get(uid, []),
-                "test_count": len(tests),
-                "test_names": tests,
-                "has_description": bool(node.get("description")),
-                "columns_documented": cols_documented,
-                "columns_total": len(columns),
-            }
-        )
-
-    # Build node list — sources
-    for uid, src in manifest.get("sources", {}).items():
-        columns = src.get("columns", {})
-        cols_documented = sum(1 for c in columns.values() if c.get("description"))
-        tests = test_map.get(uid, [])
-        nodes_out.append(
-            {
-                "id": uid,
-                "name": src.get("name", ""),
-                "type": "source",
-                "source_name": src.get("source_name", ""),
-                "schema": src.get("schema", ""),
-                "materialization": None,
-                "description": src.get("description", ""),
-                "depends_on": [],
-                "depended_on_by": depended_on_by.get(uid, []),
-                "test_count": len(tests),
-                "test_names": tests,
-                "has_description": bool(src.get("description")),
-                "columns_documented": cols_documented,
-                "columns_total": len(columns),
-            }
-        )
-
-    return {"nodes": nodes_out, "edges": edges}
-
-
-def _get_impact_tree(
-    reverse_map: dict[str, list[str]],
-    node_id: str,
-    all_nodes: dict[str, dict[str, Any]],
-    ancestors: frozenset[str] | None = None,
-) -> dict[str, Any]:
-    """Recursively build downstream impact tree from a node.
-
-    Uses an ancestor chain (not a shared visited set) so diamond dependencies
-    expand fully under each branch while true back-edge cycles are detected.
-
-    Args:
-        reverse_map: Mapping of node_id → list of downstream node_ids.
-        node_id: Starting node unique_id.
-        all_nodes: Combined dict of all manifest nodes + sources.
-        ancestors: Immutable set of ancestor node_ids for cycle detection.
-
-    Returns:
-        Tree dict with ``name``, ``type``, ``children``,
-        ``total_downstream_count``, and optionally ``cycle``.
-    """
-    if ancestors is None:
-        ancestors = frozenset()
-
-    node = all_nodes.get(node_id, {})
-    if node_id in ancestors:
-        return {
-            "name": node.get("name", node_id),
-            "type": node.get("resource_type", "unknown"),
-            "cycle": True,
-            "children": [],
-            "total_downstream_count": 0,
-        }
-
-    new_ancestors = ancestors | {node_id}
-    children = []
-    total = 0
-    for child_id in reverse_map.get(node_id, []):
-        child_tree = _get_impact_tree(reverse_map, child_id, all_nodes, new_ancestors)
-        children.append(child_tree)
-        total += 1 + child_tree["total_downstream_count"]
-
-    return {
-        "name": node.get("name", node_id),
-        "type": node.get("resource_type", "unknown"),
-        "children": children,
-        "total_downstream_count": total,
-    }
-
-
-def _build_impact_response(
-    manifest: dict[str, Any],
-    model_name: str,
-) -> dict[str, Any] | None:
-    """Build impact analysis response from manifest (blocking helper).
-
-    Searches models first, then sources, so models take priority when
-    names collide (common in dbt: source "orders" + model "orders").
-
-    Args:
-        manifest: Parsed dbt manifest dict.
-        model_name: Validated model/source name to look up.
-
-    Returns:
-        Impact response dict, or ``None`` if *model_name* not found.
-    """
-    all_nodes: dict[str, dict[str, Any]] = {}
-    reverse_map: dict[str, list[str]] = {}
-
-    for uid, node in manifest.get("nodes", {}).items():
-        all_nodes[uid] = node
-        if node.get("resource_type") == "model":
-            for dep in node.get("depends_on", {}).get("nodes", []):
-                reverse_map.setdefault(dep, []).append(uid)
-
-    for uid, src in manifest.get("sources", {}).items():
-        all_nodes[uid] = src
-
-    # Find target — prefer models over sources for name collisions
-    target_id: str | None = None
-    for uid, node in manifest.get("nodes", {}).items():
-        if node.get("resource_type") == "model" and node.get("name") == model_name:
-            target_id = uid
-            break
-    if target_id is None:
-        for uid in manifest.get("sources", {}):
-            if all_nodes[uid].get("name") == model_name:
-                target_id = uid
-                break
-
-    if target_id is None:
-        return None
-
-    tree = _get_impact_tree(reverse_map, target_id, all_nodes)
-    direct = reverse_map.get(target_id, [])
-    node_info = all_nodes[target_id]
-
-    return {
-        "model": model_name,
-        "type": node_info.get("resource_type", "unknown"),
-        "direct_dependents": [all_nodes.get(d, {}).get("name", d) for d in direct],
-        "tree": tree,
-        "total_downstream_count": tree["total_downstream_count"],
-    }
 
 
 # ---------------------------------------------------------------------------
